# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Meridian Commodities internal trade analytics API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface Shipment {
  id: string;
  shipmentDate: string;
  customer: string;
  country: string;
  quantityMt: number;
  sellingPriceFob: number;
  purchaseCost: number;
  freightInsurance: number;
  /** @nullable */
  vesselName: string | null;
  /** @nullable */
  product: string | null;
  /** @nullable */
  loadingCompany: string | null;
  /** @nullable */
  loadingPort: string | null;
  createdAt: string;
  revenue: number;
  cost: number;
  profit: number;
  margin: number;
}

export interface ShipmentInput {
  shipmentDate: string;
  /** @minLength 1 */
  customer: string;
  /** @minLength 1 */
  country: string;
  /** @minimum 0 */
  quantityMt: number;
  /** @minimum 0 */
  sellingPriceFob?: number;
  /** @minimum 0 */
  purchaseCost?: number;
  /** @minimum 0 */
  freightInsurance?: number;
  vesselName?: string;
  product?: string;
  loadingCompany?: string;
  loadingPort?: string;
}

export interface ShipmentUpdate {
  shipmentDate?: string;
  /** @minLength 1 */
  customer?: string;
  /** @minLength 1 */
  country?: string;
  /** @minimum 0 */
  quantityMt?: number;
  /** @minimum 0 */
  sellingPriceFob?: number;
  /** @minimum 0 */
  purchaseCost?: number;
  /** @minimum 0 */
  freightInsurance?: number;
  vesselName?: string;
  product?: string;
  loadingCompany?: string;
  loadingPort?: string;
}

export interface ShipmentList {
  items: Shipment[];
  total: number;
  page: number;
  pageSize: number;
  hasMore: boolean;
}

export interface Metrics {
  revenue: number;
  profit: number;
  volumeMt: number;
  margin: number;
  shipmentCount: number;
}

export type CustomerSummary = Metrics & {
  customer: string;
};

export type CountrySummary = Metrics & {
  country: string;
  customers: CustomerSummary[];
};

export interface DashboardSummary {
  metrics: Metrics;
  countries: CountrySummary[];
}

export type SupplierMarket = Metrics & {
  country: string;
};

export type SupplierSummaryItem = Metrics & {
  loadingCompany: string;
  markets: SupplierMarket[];
};

export interface SupplierSummary {
  metrics: Metrics;
  suppliers: SupplierSummaryItem[];
}

export interface Metadata {
  customers: string[];
  countries: string[];
  products: string[];
  loadingCompanies: string[];
  loadingPorts: string[];
  years: number[];
}

export interface ResetResult {
  count: number;
}

export interface ImportResult {
  count: number;
}

/**
 * Invalid request
 */
export type BadRequestResponse = Error;

/**
 * Not found
 */
export type NotFoundResponse = Error;

export type PageParameter = number;

export type PageSizeParameter = number;

export type SearchParameter = string;

export type YearParameter = number;

export type ListShipmentsParams = {
/**
 * @minimum 1
 */
page?: PageParameter;
/**
 * @minimum 1
 * @maximum 100
 */
pageSize?: PageSizeParameter;
search?: SearchParameter;
/**
 * @minimum 2020
 * @maximum 2100
 */
year?: YearParameter;
country?: string[];
product?: string;
loadingCompany?: string;
from?: string;
to?: string;
};

export type GetDashboardSummaryParams = {
/**
 * @minimum 2020
 * @maximum 2100
 */
year?: YearParameter;
country?: string[];
product?: string;
loadingCompany?: string;
from?: string;
to?: string;
};