# lib/api-zod/src/generated/types/countrySummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Meridian Commodities internal trade analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { CustomerSummary } from './customerSummary';
import type { Metrics } from './metrics';

export type CountrySummary = Metrics & {
  country: string;
  customers: CustomerSummary[];
};