# lib/api-zod/src/generated/types/customerSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Meridian Commodities internal trade analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { Metrics } from './metrics';

export type CustomerSummary = Metrics & {
  customer: string;
};