# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Meridian Commodities internal trade analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { CountrySummary } from './countrySummary';
import type { Metrics } from './metrics';

export interface DashboardSummary {
  metrics: Metrics;
  countries: CountrySummary[];
}