# lib/api-zod/src/generated/types/resetResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Meridian Commodities internal trade analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface ResetResult {
  count: number;
}